Add logic of the Buffer in base.py and add regressor and classifier as API.
